+++ crates/forge-runner/src/lib.rs
@@ -13,11 +13,13 @@ use cairo_lang_sierra_to_casm::compiler::CairoProgram;
 use cairo_lang_sierra_to_casm::metadata::{calc_metadata, MetadataComputationConfig};
 use camino::Utf8PathBuf;
 
+use contracts_data::ContractsData;
 use futures::stream::FuturesUnordered;
 use futures::StreamExt;
 
 use once_cell::sync::Lazy;
 use scarb_api::StarknetContractArtifacts;
+use running::TestDetails;
 use smol_str::SmolStr;
 use trace_data::save_trace_data;
 
@@ -29,6 +31,7 @@ use tokio::sync::mpsc::{channel, Sender};
 use tokio::task::JoinHandle;
 
 pub mod compiled_runnable;
+pub mod contracts_data;
 pub mod expected_result;
 pub mod test_case_summary;
 pub mod test_crate_summary;
@@ -91,18 +94,18 @@ impl RunnerConfig {
 #[non_exhaustive]
 #[derive(Debug, Clone)]
 pub struct RunnerParams {
-    contracts: HashMap<String, StarknetContractArtifacts>,
+    contracts_data: ContractsData,
     environment_variables: HashMap<String, String>,
 }
 
 impl RunnerParams {
     #[must_use]
     pub fn new(
-        contracts: HashMap<String, StarknetContractArtifacts>,
+        contracts_data: ContractsData,
         environment_variables: HashMap<String, String>,
     ) -> Self {
         Self {
-            contracts,
+            contracts_data,
             environment_variables,
         }
     }