+++ crates/forge/tests/e2e/mod.rs
@@ -10,7 +10,7 @@ mod forking;
 mod fuzzing;
 mod io_operations;
 mod running;
-mod trace;
 mod trace_data;
+mod trace_print;
 mod trace_resources;
 mod workspaces;